Extract common GLSL loader code to allow for separate webpack and vite loaders
